refactor: 댓글 목록 컴포넌트 UserCommentList로 분리 및 스타일 추가
